Split map feature model tests and fakes #294
